updated landing page designs
