+++ packages/react/src/Hero/Hero.module.css
@@ -51,7 +51,7 @@
 
 .Hero-heading {
   margin: 0;
-  max-width: 24ch;
+  max-width: 18ch;
 }
 
 .Hero-heading--fullWidth {
