+++ packages/react/src/recipes/SolutionTemplates/SolutionPage/SolutionPage.module.css
@@ -1,0 +1,259 @@
+.hideHorizontalOverflow {
+  overflow-x: hidden;
+}
+
+.sectionIntro {
+  display: grid;
+  grid-template-rows: auto;
+  height: 100%;
+  grid-auto-rows: max-content;
+  justify-items: start;
+}
+
+.verticalOffset > div:nth-child(2),
+.verticalOffset > div:nth-child(3) {
+  margin-block-start: calc(var(--base-size-48) * -1);
+}
+
+.topOfStack {
+  position: relative;
+  z-index: 2;
+}
+
+.relative {
+  position: relative;
+  z-index: 1;
+}
+
+.gitLinesBg {
+  background-size: cover;
+  background-position: center center;
+}
+
+.ctaBanner {
+  --brand-CTABanner-bgColor: var(--brand-color-canvas-subtle);
+}
+
+.noBottomRadius {
+  border-bottom-left-radius: 0;
+  border-bottom-right-radius: 0;
+}
+
+.jtbd2Section {
+  --brand-RiverStoryScroll-content-inactive-bgColor: var(--brand-color-canvas-subtle);
+}
+
+/*
+ * Hero video glow
+ * Opt-in feature. Not officialy-support Primer Brand pattern.
+ */
+
+@keyframes scaleAnimation {
+  0% {
+    opacity: 0.1;
+    transform: translateX(200px) scale(0.5);
+  }
+  100% {
+    opacity: 1;
+    transform: translateX(200px) scale(1);
+  }
+}
+
+.heroVideoGlow {
+  position: relative;
+}
+
+.heroVideoGlow::before,
+.heroVideoGlow::after {
+  --diameter: 520px;
+  --radius: 720px;
+  --topOffset: 600px;
+  --leftOffset: 45%;
+
+  position: absolute;
+  top: var(--topOffset);
+  bottom: 0;
+  left: var(--leftOffset);
+  z-index: -1;
+  display: block;
+  width: var(--diameter);
+  height: var(--diameter);
+  transform: translateX(200px) scale(0.99);
+  transform-origin: center 0;
+  content: '';
+  z-index: 0;
+  filter: blur(200px);
+  animation: scaleAnimation 3s var(--brand-animation-easing-default) forwards;
+  animation-delay: 0.3s;
+  opacity: 0;
+  border-radius: var(--radius);
+}
+
+.heroVideoGlow::before {
+  left: calc(var(--leftOffset) - 200px);
+  top: calc(--topOffset - 100px);
+}
+
+[data-color-mode='light'] .heroVideoGlow::before {
+  background: var(--base-color-scale-blue-2);
+}
+
+[data-color-mode='light'] .heroVideoGlow::after {
+  background: var(--base-color-scale-purple-2);
+}
+
+[data-color-mode='dark'] .heroVideoGlow::before {
+  background: var(--base-color-scale-blue-7);
+}
+
+[data-color-mode='dark'] .heroVideoGlow::after {
+  background: var(--base-color-scale-purple-7);
+}
+
+/*
+ * Parallax effect
+ * Opt-in feature. Not officialy-support Primer Brand pattern.
+ */
+
+.hero {
+  --parallax-start-spacing: 74px;
+  --hero-height: 90vh;
+
+  position: relative;
+  height: calc(var(--hero-height) - var(--parallax-start-spacing));
+}
+
+@media screen and (min-height: 700px) {
+  .hero {
+    --hero-height: 80vh;
+  }
+}
+
+@media screen and (min-height: 900px) {
+  .hero {
+    --hero-height: 75vh;
+  }
+}
+
+@media screen and (min-height: 1000px) {
+  .hero {
+    --hero-height: 60vh;
+  }
+}
+
+.parallax {
+  position: absolute;
+  top: calc(-1 * var(--parallax-start-spacing));
+  width: 100%;
+  height: var(--hero-height);
+  max-height: 600px;
+}
+
+.background,
+.foreground {
+  width: 100%;
+  height: 100%;
+
+  animation-fill-mode: forwards;
+  animation-timing-function: linear;
+  animation-timeline: scroll();
+  animation-range: 0vh var(--hero-height);
+}
+
+.background {
+  position: absolute;
+}
+
+.foreground {
+  position: relative;
+}
+
+.heroImage {
+  min-width: 100%;
+  min-height: 100%;
+}
+
+.heroImageOverlay {
+  position: absolute;
+  top: 0;
+  width: 100%;
+  height: 100%;
+}
+
+.lightHero .heroImageOverlay {
+  background: linear-gradient(
+    0deg,
+    rgba(255, 255, 255, 0.8) 0%,
+    rgba(255, 255, 255, 0) 50%,
+    rgba(255, 255, 255, 0.6) 100%
+  );
+}
+
+.darkHero .heroImageOverlay {
+  background: linear-gradient(0deg, rgba(0, 0, 0, 0.8) 0%, rgba(0, 0, 0, 0) 50%, rgba(0, 0, 0, 0.6) 100%);
+}
+
+.heroImageOverlayDark {
+  position: absolute;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(0deg, rgba(0, 0, 0, 0.8) 0%, rgba(0, 0, 0, 0) 50%, rgba(0, 0, 0, 0.6) 100%);
+}
+
+.foregroundContents {
+  padding: 0;
+  padding-block-start: calc(var(--base-size-20) + var(--parallax-start-spacing));
+  height: 100%;
+}
+
+.articleContents {
+  border-radius: var(--brand-borderRadius-xlarge);
+  margin-top: calc(-1 * var(--brand-borderRadius-xlarge));
+  background-color: var(--brand-color-canvas-default);
+  position: relative;
+  z-index: 90;
+}
+
+.lightHero .articleContents {
+  box-shadow: 0px -2px 49px 0px rgba(0, 0, 0, 0.25);
+}
+
+@keyframes parallax-background {
+  from {
+    top: 0;
+    filter: blur(0px);
+  }
+  to {
+    top: calc(var(--hero-height) * -0.4);
+    filter: blur(0) brightness(0%);
+  }
+}
+
+@keyframes parallax-foreground {
+  from {
+    top: 0;
+  }
+  to {
+    top: calc(calc(var(--hero-height) * 2) * -0.2);
+    filter: blur(10px) brightness(0%);
+  }
+}
+
+/* Progressive enhancement opt-in */
+@supports (animation-timeline: scroll()) {
+  @media (prefers-reduced-motion: no-preference) {
+    .parallax {
+      position: fixed;
+      top: 0;
+    }
+
+    .foreground {
+      animation-name: parallax-foreground;
+    }
+
+    .background {
+      animation-name: parallax-background;
+    }
+  }
+}
