+++ packages/react/src/recipes/SolutionTemplates/SolutionPage/SolutionPage.module.css.d.ts
@@ -1,0 +1,28 @@
+declare const styles: {
+  readonly "hideHorizontalOverflow": string;
+  readonly "sectionIntro": string;
+  readonly "verticalOffset": string;
+  readonly "topOfStack": string;
+  readonly "relative": string;
+  readonly "gitLinesBg": string;
+  readonly "ctaBanner": string;
+  readonly "noBottomRadius": string;
+  readonly "jtbd2Section": string;
+  readonly "heroVideoGlow": string;
+  readonly "scaleAnimation": string;
+  readonly "hero": string;
+  readonly "parallax": string;
+  readonly "background": string;
+  readonly "foreground": string;
+  readonly "heroImage": string;
+  readonly "heroImageOverlay": string;
+  readonly "lightHero": string;
+  readonly "darkHero": string;
+  readonly "heroImageOverlayDark": string;
+  readonly "foregroundContents": string;
+  readonly "articleContents": string;
+  readonly "parallax-foreground": string;
+  readonly "parallax-background": string;
+};
+export = styles;
+
