+++ packages/react/src/recipes/SolutionTemplates/SolutionPage/UseCase/UseCase.stories.tsx
@@ -2,13 +2,15 @@ import React from 'react'
 import {INITIAL_VIEWPORTS} from '@storybook/addon-viewport'
 import {Meta, StoryFn} from '@storybook/react'
 
-import {SolutionPage} from './SolutionPage'
-import {sharedArgTypes} from '../helpers'
+import {SolutionPage} from '../SolutionPage'
+import {sharedArgTypes} from '../../helpers'
+import {ColorModesEnum} from '../../../../ThemeProvider'
 
 export default {
   title: 'Recipes/Solutions/Solution: Use case',
   component: SolutionPage,
   args: {
+    variant: 'use-case',
     heroImage: true,
     heroLabel: 'DevOps',
     heroTitle: 'The complete CI/CD solution',
@@ -16,6 +18,8 @@ export default {
       'Build, test, and deploy software with simple and secure enterprise CI/CD, all on the complete development platform.',
     heroCtaTextPrimary: 'Start a free trial',
     heroCtaTextSecondary: 'Contact Sales',
+    sectionIntroText: 'A single, integrated, enterprise-ready platform',
+    sectionIntroCTAText: 'Explore GitHub Enterprise',
     introVariant: 'editorial list',
 
     logoBarVisible: true,
@@ -36,40 +40,40 @@ export default {
   },
 } as Meta<typeof SolutionPage>
 
-export const Playground: StoryFn<typeof SolutionPage> = args => <SolutionPage {...args} />
-Playground.args = {
+const maximumArgs = {
   variant: 'use-case',
-
-  sectionIntroText: 'A single, integrated, enterprise-ready platform',
-  sectionIntroCTAText: 'Explore GitHub Enterprise',
-
   logoBarVisible: true,
   riverVisible: true,
   customerStoryVisible: false,
   testimonialsVisible: true,
   faqVisible: true,
   statisticsVisible: true,
+}
+export const Maximum: StoryFn<typeof SolutionPage> = args => <SolutionPage {...args} />
+Maximum.args = {
+  ...maximumArgs,
+  variant: 'use-case',
+}
+Maximum.storyName = 'Maximum Light'
+
+export const MaximumDark: StoryFn<typeof SolutionPage> = args => <Maximum {...args} />
+MaximumDark.args = {
+  ...maximumArgs,
+  variant: 'use-case',
+  colorMode: ColorModesEnum.DARK,
 }
 
 export const Minimum: StoryFn<typeof SolutionPage> = args => <SolutionPage {...args} />
 Minimum.args = {
   variant: 'use-case',
-
-  sectionIntroText: 'A single, integrated, enterprise-ready platform',
-  sectionIntroCTAText: 'Explore GitHub Enterprise',
-}
-
-export const Maximum: StoryFn<typeof SolutionPage> = args => <SolutionPage {...args} />
-Maximum.args = {
-  variant: 'use-case',
-
-  sectionIntroText: 'A single, integrated, enterprise-ready platform',
-  sectionIntroCTAText: 'Explore GitHub Enterprise',
-
-  logoBarVisible: true,
   riverVisible: true,
-  customerStoryVisible: false,
-  testimonialsVisible: true,
-  faqVisible: true,
-  statisticsVisible: true,
 }
+Minimum.storyName = 'Minimum Light'
+
+export const MinimumDark = Minimum.bind({})
+MinimumDark.args = {
+  variant: 'use-case',
+  colorMode: ColorModesEnum.DARK,
+  riverVisible: true,
+}
+MinimumDark.storyName = 'Minimum Dark'
