+++ packages/react/src/recipes/SolutionTemplates/SolutionPage/UseCase/UseCase.visual.spec.ts
@@ -1,0 +1,45 @@
+/*
+ * Do not modify this file directly.
+ * This file was generated by: playwright.generate-tests.ts.
+ * Regenerate using: npm run test:visual:generate
+ */
+import {test, expect} from '@playwright/test'
+
+// eslint-disable-next-line i18n-text/no-en
+test.describe('Visual Comparison: UseCase', () => {
+  test('UseCase / Maximum Light', async ({page}) => {
+    await page.goto(
+      'http://localhost:6006/iframe.html?args=&id=recipes-solutions-solution-use-case--maximum&viewMode=story',
+    )
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  test('UseCase / Maximum Dark', async ({page}) => {
+    await page.goto(
+      'http://localhost:6006/iframe.html?args=&id=recipes-solutions-solution-use-case--maximum-dark&viewMode=story',
+    )
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  test('UseCase / Minimum Light', async ({page}) => {
+    await page.goto(
+      'http://localhost:6006/iframe.html?args=&id=recipes-solutions-solution-use-case--minimum&viewMode=story',
+    )
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  test('UseCase / Minimum Dark', async ({page}) => {
+    await page.goto(
+      'http://localhost:6006/iframe.html?args=&id=recipes-solutions-solution-use-case--minimum-dark&viewMode=story',
+    )
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+})
