+++ packages/react/src/recipes/SolutionTemplates/SolutionPage/index.ts
