+++ packages/react/src/recipes/SolutionTemplates/SolutionPortal/SolutionPortal.visual.spec.ts
@@ -1,0 +1,23 @@
+/*
+ * Do not modify this file directly.
+ * This file was generated by: playwright.generate-tests.ts.
+ * Regenerate using: npm run test:visual:generate
+ */
+import {test, expect} from '@playwright/test'
+
+// eslint-disable-next-line i18n-text/no-en
+test.describe('Visual Comparison: SolutionPortal', () => {
+  test('SolutionPortal / Light', async ({page}) => {
+    await page.goto('http://localhost:6006/iframe.html?args=&id=recipes-solutions-portal--light&viewMode=story')
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  test('SolutionPortal / Dark', async ({page}) => {
+    await page.goto('http://localhost:6006/iframe.html?args=&id=recipes-solutions-portal--dark&viewMode=story')
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+})
