--- packages/react/src/recipes/SolutionTemplates/SolutionTemplate/SolutionPage.module.css
@@ -1,93 +1,0 @@
-.sectionIntro {
-  display: grid;
-  grid-template-rows: auto;
-  height: 100%;
-  grid-auto-rows: max-content;
-  justify-items: start;
-}
-
-.verticalOffset > div:nth-child(2),
-.verticalOffset > div:nth-child(3) {
-  margin-block-start: calc(var(--base-size-48) * -1);
-}
-
-.topOfStack {
-  position: relative;
-  z-index: 2;
-}
-
-@keyframes scaleAnimation {
-  0% {
-    transform: translateY(-10%) scale(0);
-    opacity: 0;
-  }
-  100% {
-    transform: translateY(0) scale(1);
-  }
-}
-
-.customHero {
-  position: relative;
-}
-
-.customHero::before,
-.customHero::after {
-  content: '';
-  position: absolute;
-  top: 390px;
-  right: -15%;
-  z-index: -1;
-  display: block;
-  width: 1000px;
-  height: 1000px;
-  transform-origin: center center;
-  z-index: 0;
-  transform: translateY(-10%) scale(0);
-  animation: scaleAnimation 4s var(--brand-animation-easing-default) forwards;
-  animation-delay: 1s;
-  opacity: 0;
-}
-
-.customHero::before {
-  top: 420px;
-  right: -20%;
-}
-
-[data-color-mode='light'] .customHero::before {
-  background: radial-gradient(50% 50% at 50% 50%, var(--base-color-scale-blue-2) 0, rgba(195, 186, 250, 0) 100%);
-  opacity: 0.6;
-}
-
-[data-color-mode='light'] .customHero::after {
-  background: radial-gradient(50% 50% at 50% 50%, var(--base-color-scale-purple-2) 0, rgba(195, 186, 250, 0) 100%);
-  opacity: 0.9;
-}
-
-[data-color-mode='dark'] .customHero::before {
-  background: radial-gradient(50% 50% at 50% 50%, var(--base-color-scale-blue-8) 0, rgba(195, 186, 250, 0) 100%);
-  opacity: 0.53;
-}
-
-[data-color-mode='dark'] .customHero::after {
-  background: radial-gradient(50% 50% at 50% 50%, var(--base-color-scale-purple-8) 0, rgba(195, 186, 250, 0) 100%);
-  opacity: 0.9;
-}
-
-.relative {
-  position: relative;
-  z-index: 1;
-}
-
-.gitLinesBg {
-  background-size: cover;
-  background-position: center center;
-}
-
-.ctaBanner {
-  --brand-CTABanner-bgColor: var(--brand-color-canvas-subtle);
-}
-
-.noBottomRadius {
-  border-bottom-left-radius: 0;
-  border-bottom-right-radius: 0;
-}
