--- packages/react/src/recipes/SolutionTemplates/SolutionTemplate/SolutionPage.module.css.d.ts
@@ -1,13 +1,0 @@
-declare const styles: {
-  readonly "sectionIntro": string;
-  readonly "verticalOffset": string;
-  readonly "topOfStack": string;
-  readonly "customHero": string;
-  readonly "scaleAnimation": string;
-  readonly "relative": string;
-  readonly "gitLinesBg": string;
-  readonly "ctaBanner": string;
-  readonly "noBottomRadius": string;
-};
-export = styles;
-
