+++ packages/react/src/river/RiverStoryScroll/RiverStoryScroll.visual.spec.ts
@@ -1,0 +1,88 @@
+/*
+ * Do not modify this file directly.
+ * This file was generated by: playwright.generate-tests.ts.
+ * Regenerate using: npm run test:visual:generate
+ */
+import {test, expect} from '@playwright/test'
+
+// eslint-disable-next-line i18n-text/no-en
+test.describe('Visual Comparison: RiverStoryScroll', () => {
+  test('RiverStoryScroll / With Timeline', async ({page}) => {
+    await page.goto(
+      'http://localhost:6006/iframe.html?args=&id=components-riverstoryscroll-features--with-timeline&viewMode=story',
+    )
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  // eslint-disable-next-line i18n-text/no-en
+  test.describe('Mobile viewport test for With Timeline Narrow', () => {
+    test.use({viewport: {width: 360, height: 800}})
+    test('RiverStoryScroll / With Timeline Narrow', async ({page}) => {
+      await page.goto(
+        'http://localhost:6006/iframe.html?args=&id=components-riverstoryscroll-features--with-timeline-narrow&viewMode=story',
+      )
+
+      await page.waitForTimeout(500)
+      expect(await page.screenshot()).toMatchSnapshot()
+    })
+  })
+  test('RiverStoryScroll / Disable', async ({page}) => {
+    await page.goto(
+      'http://localhost:6006/iframe.html?args=&id=components-riverstoryscroll-features--disable&viewMode=story',
+    )
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  test('RiverStoryScroll / Enterprise Example', async ({page}) => {
+    await page.goto(
+      'http://localhost:6006/iframe.html?args=&id=components-riverstoryscroll-features--enterprise-example&viewMode=story',
+    )
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  // eslint-disable-next-line i18n-text/no-en
+  test.describe('Mobile viewport test for Enterprise Example Narrow', () => {
+    test.use({viewport: {width: 360, height: 800}})
+    test('RiverStoryScroll / Enterprise Example Narrow', async ({page}) => {
+      await page.goto(
+        'http://localhost:6006/iframe.html?args=&id=components-riverstoryscroll-features--enterprise-example-narrow&viewMode=story',
+      )
+
+      await page.waitForTimeout(500)
+      expect(await page.screenshot()).toMatchSnapshot()
+    })
+  })
+  test('RiverStoryScroll / Video', async ({page}) => {
+    await page.goto(
+      'http://localhost:6006/iframe.html?args=&id=components-riverstoryscroll-features--video&viewMode=story',
+    )
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+
+  // eslint-disable-next-line i18n-text/no-en
+  test.describe('Mobile viewport test for Video Narrow', () => {
+    test.use({viewport: {width: 360, height: 800}})
+    test('RiverStoryScroll / Video Narrow', async ({page}) => {
+      await page.goto(
+        'http://localhost:6006/iframe.html?args=&id=components-riverstoryscroll-features--video-narrow&viewMode=story',
+      )
+
+      await page.waitForTimeout(500)
+      expect(await page.screenshot()).toMatchSnapshot()
+    })
+  })
+  test('RiverStoryScroll / Default', async ({page}) => {
+    await page.goto('http://localhost:6006/iframe.html?args=&id=components-riverstoryscroll--default&viewMode=story')
+
+    await page.waitForTimeout(500)
+    expect(await page.screenshot()).toMatchSnapshot()
+  })
+})
